requirements.txt
asgiref==3.4.1
bandit==1.4.0
black==22.3.0
click==8.0.4
Django==3.2.14
djangorestframework==3.13.1
drf-yasg==1.21.0
importlib-metadata==4.8.3
isort==5.10.1
mypy-extensions==0.4.3
pathspec==0.9.0
prospector==0.12.7
platformdirs==2.4.0
pylint==1.8.3
pytz==2022.1
sqlparse==0.4.2
tomli==1.2.3
typed-ast==1.5.4
typing-extensions==4.1.1
zipp==3.6.0